Updated Signup Module
In this commit I updated the signup module in such a way that first it checks if the details of students, faculty members and industrial professionals are given in strings that first it parses into JSON objects and then stores it. Also I made concurrency on the frontend and backend for roles of students, faculties and professionals.
Now it is working as expected.